moving lots of stuff around

+++ beamertoy/include/Shadertoy.h
@@ -47,9 +47,9 @@ namespace shadertoy {
         int id;
         std::string src;  //url or path
         std::string ctype; //buffer cubemap keyboard mic music musicstream texture video volume webcam
-        int64_t channel; // channel number, [0,1,2,3]
+        size_t channel; // channel number, [0,1,2,3]
         struct Sampler sampler;
-        int64_t published;
+        size_t published;
 
         template<class Archive>
         void serialize(Archive &archive) {
